==> src/recipients/1.parse_mmvec.ipynb <==
import pandas as pd
import numpy as np
from skbio.stats.composition import clr_inv as softmax
import warnings
import os
os.chdir("../../data/recipients/")
warnings.filterwarnings("ignore")

# Recipient

def convert_df(df):
    df.columns=["HEAD","TAIL","WEIGHT"]
    df=df.sort_values('WEIGHT', ascending=False)
    df=df.iloc[0:int(len(df)*0.10),]
    return df

ranks = pd.read_table('mmvec_recipients/latent_dim_3_input_prior_1.00_output_prior_1.00_beta1_0.90_beta2_0.95_ranks.txt')
df = pd.melt(ranks, id_vars='featureid', value_vars=ranks.columns,ignore_index=True)


df_hx=df[df["featureid"].str.contains('_hx')]
df_fx=df[df["featureid"].str.contains('_fx')]
df_feces=df[df["featureid"].str.contains('_feces')]
df_serum=df[df["featureid"].str.contains('_serum')]
df_meta=df[~df["featureid"].str.contains('_serum|_hx|_fx|_feces')]

df_feces=convert_df(df_feces)
df_serum=convert_df(df_serum)
df_fx=convert_df(df_fx)
df_hx=convert_df(df_hx)
df_meta=convert_df(df_meta)

df = pd.concat([df_meta,df_feces,df_serum,df_fx,df_hx])

df.to_csv("mmvec.edge.recipients.txt",sep="\t",index=False)
