==> src/recipients/2.parse_wgcna_modules.ipynb <==
# Coded in R

library(WGCNA)
library(dplyr)
setwd("../../data/recipients/")

getrows<-function(modules,path){
    rows<-c()
    for(i in modules){
        module.file<-paste0(path,i,".txt")
        module<-read.csv(module.file,header=F,sep="\t")
        rows<-c(rows,module$V1)
    }
    return(rows)
}
add_data_type_to_mtx<-function(df,type){
    idx=c()
    for (i in rownames(df)){
        idx=c(idx,paste0(i,"_",type))
    }
    rownames(df)<-idx
    return(df)
}
#wKDA prep
getmodulesrows<-function(modules,path,type){
    module<-data.frame()
    rows<-c()
    df <- data.frame(matrix(ncol = 2, nrow = 0))
    colnames(df) <- c('MODULE', 'NODE')
    for(i in modules){
        module.file<-paste0(path,i,".txt")
        module<-read.csv(module.file,header=F,sep="\t")
        colnames(module)<-c("NODE")
        module$MODULE=paste0(i,"_",type)
        df<-rbind(df,module)
    }
    
    return(df)
}

selected.modules<-read.csv("selected_modules.tsv",sep="\t")

meta.modules<-c("red","lightcyan")
feces.modules<-c("grey60","darkgrey")
serum.modules<-c("darkgrey","green","cyan")
hx.modules<-c("brown","ivory","lightcyan1","grey60","brown4","salmon","red","lightpink4")
fx.modules<-c("mediumpurple2","darkseagreen4","lavenderblush3","mediumpurple3","antiquewhite4","plum")

data.meta<-read.csv("raw_data/metagenome_new.tsv",sep="\t",row.names = 1)
data.meta<-log(data.meta+1)
data.feces<-read.csv("raw_data/metab_feces_norm.csv",row.names=1)
data.serum<-read.csv("raw_data/metab_serum_norm.csv",row.names=1)
data.fx<-read.csv("raw_data/fx_PrePost_RNAseq_counts.log.csv",row.names=1)
data.hx<-read.csv("raw_data/hx_PrePost_RNAseq_counts.log.csv",row.names=1)

row.meta<-getrows(meta.modules,"wgcna/modules/metagenome_modules/")
row.feces<-getrows(feces.modules,"wgcna/modules/modules-feces/modules/")
row.serum<-getrows(serum.modules,"wgcna/modules/modules-serum/modules/")
row.fx<-getrows(fx.modules,"wgcna/modules/modules-fx/original/")
row.hx<-getrows(hx.modules,"wgcna/modules/modules-hx/original/")

data.meta<-data.meta[row.meta,]
data.feces<-data.feces[row.feces,]
data.serum<-data.serum[row.serum,]
data.fx<-data.fx[row.fx,]
data.hx<-data.hx[row.hx,]

data.feces<-add_data_type_to_mtx(data.feces,"feces")
data.serum<-add_data_type_to_mtx(data.serum,"serum")
data.fx<-add_data_type_to_mtx(data.fx,"fx")
data.hx<-add_data_type_to_mtx(data.hx,"hx")

data<-rbind(data.meta,data.feces,data.serum,data.fx,data.hx)
data<-t(data)

#DEGs and GWAS added manually after this step, and then wKDA
mod.meta<-getmodulesrows(meta.modules,"wgcna/modules/metagenome_modules/","meta")
mod.feces<-getmodulesrows(feces.modules,"wgcna/modules/modules-feces/modules/","feces")
mod.serum<-getmodulesrows(serum.modules,"wgcna/modules/modules-serum/modules/","serum")
mod.fx<-getmodulesrows(fx.modules,"wgcna/modules/modules-fx/original/","fx")
mod.hx<-getmodulesrows(hx.modules,"wgcna/modules/modules-hx/original/","hx")


degs.hx<-read.csv("DEGs/Degs_hippo.txt")
degs.fx<-read.csv("DEGs/Degs_frontal_cortex.txt")
en2gene.fx<-read.csv("gene2ensemble_fx.csv")
en2gene.hx<-read.csv("gene2ensemble_hx.csv")
ensmbl.fx<-c()
for (i in degs.fx$DEGS){
   ensmbl.fx<-c(ensmbl.fx,en2gene.fx[en2gene.fx$gene==grep(paste0("^",i,"$"),en2gene.fx$gene,ignore.case=TRUE,value=TRUE),]$ensbl)
}
degs.fx$DEGS=paste0(ensmbl.fx,"_fx")
colnames(degs.fx)=c("NODE")
degs.fx$MODULE<-"DEG_fx"
ensmbl.hx<-c()
for (i in degs.hx$DEGS){
   ensmbl.hx<-c(ensmbl.hx,en2gene.hx[en2gene.hx$gene==grep(paste0("^",i,"$"),en2gene.hx$gene,ignore.case=TRUE,value=TRUE),]$ensbl)
}
degs.hx$DEGS=paste0(ensmbl.hx,"_fx")
colnames(degs.hx)=c("NODE")
degs.hx$MODULE<-"DEG_hx"

dam.feces<-read.csv("dam_feces.txt",sep="\t")
dam.feces$MODULE="DEG_feces"
dam.feces$NODE<-paste0(dam.feces$NODE,"_feces")
dam.serum<-read.csv("dam_serum.txt",sep="\t")
dam.serum$MODULE="DEG_serum"
dam.serum$NODE=paste0(dam.serum$NODE,"_serum")

gwas.overlap<-read.csv("GWAS.overlap.txt")
colnames(gwas.overlap)<-c("NODE")
gwas.overlap$NODE<-paste0(gwas.overlap$NODE,"_hx")
gwas.overlap$MODULE<-"GWAS_hx"

mod.fx$NODE<-paste0(mod.fx$NODE,"_fx")
mod.hx$NODE<-paste0(mod.hx$NODE,"_hx")
mod.feces$NODE<-paste0(mod.feces$NODE,"_feces")
mod.serum$NODE<-paste0(mod.serum$NODE,"_serum")

mod<-rbind(mod.meta,mod.feces,mod.serum,mod.fx,mod.hx,dam.feces,dam.serum,degs.fx,degs.hx,gwas.overlap)
write.table(mod,file="NODES4wKDA.tsv",sep="\t",quote=F,row.names = F)
